Added Exercise 7 Template
The last exercise in lab 4b is an unnecessary time sink. We provided a template that still makes them apply the averager function they made previously, know how to display things on the I2C, and have general knowledge on the pins. This template could be subject to change to make it harder on the students(right now it is a bit easy).